Separate out utility functions.
