+++ .gitignore
@@ -44,3 +44,4 @@ _map.html
 
 # Test cache
 .cache
+.pytest_cache
